minor organization
